requirements.txt
python-telegram-bot
pg8000
sqlalchemy